Ender Bucket: Fix the non-flowing water/lava bug + other fluid handling stuff.
Some fixes and clean-ups to the fluid handling code in the Ender Bucket code.
We now check if we have the vanilla water or lava source blocks, and
convert them to the flowing variant before trying to place them.
Also fixed some other fluid handling bugs, where we didn't properly check
if the fluids are identical, which allowed for duplication in some cases.