Add a "teleport self" mode to the Ender Bow. Add tp sounds/particles to items.
Added separate icons for the "tp self" mode. The Bow's mode can be toggled
with a hotkey, which is by default the middle mouse button.

Also, some items were missing teleport sounds and/or particles.
They now have them, in both the source and destination sides.
The Bow (or the arrows to be more precise) however doesn't currently
have particles, due to the way that part of the code doesn't get executed
on the client side.