Teleportation rewrites. Now allows player tp and mounted entity 'stack' tp.
Has a special case for destroying EntityMinecartContainer type entities,
since they normally drop all their contents when using removeEntoty(),
which calls setDead(), which with those drops the contents.
The overridden vanilla travelToDimension() in EntityMinecartContainer sets the
dropContentsWhenDead = false, but we can't use that method nor set the
flag because it's private. (Reflection?)