Chunkloading: Fix we didn't clean up the timeouts map when releasing tickets.
There was a crash if the server/game was shut down before the chunk timeouts
ticked down to zero. That's because we release the temporary player tickets
when a world loads and the tickets are offered back to us. But we didn't
clear the chunks belonging to those tickets from the timeouts map.
Thus when the chunks ticked down to zero and we tried to unforce them,
the game crashed because those tickets had been released already.

Now we remove the chunks from the timeouts map when ticking the chunk
timeouts, if the ticket doesn't exist in the playerTickets map anymore.