Tool Workstation: Fixed the module writing derps, hopefully.
Adding writing the modules to the tool to the getStackInSlot() method
seemed to fix it, because the transferStackInSlot() method calls that to
get the stack when it starts doing stuff. So just having it in the
setInventorySlotContents() didn't account for the transfer method having
a copy of the stack without the modules written in yet.

NOTE: has lots of debug prints, they will be removed in the next commit.