Ender Bucket: Fix a client side crash
The getCapacityCached() method did not return from the "bucket linked"
code block before reaching the regular getCapacity() call in the end. That
caused the getCapacity() method to call getLinkedTank() on the client
side, which crashes the game, at least in single player when the server
side thread modifies the same data in another thread.