Refactor + clean up some module/item stuff. Add interface for modules.
The selected module is now stored in NBT based on the name string in the enum.
Also, the GUI slot background offset is now defined in the GuiToolWorkstation
class. So the order of the module types in the enum doesn't matter anymore.

Also add an interface to modules, that allows getting the module type
and tier, and that definition is now in the item class instead
of the util class UtilItemModular.

Fix a NBT saving bug in the Ender Bag, probably introduced during
the big "clean-up" (aka. break-all-the-things).