Add compatibility target NBT data transfer from items to a link crystal.
When items are modified in the Tool Workstation, we call a method
to check if the item has old target data directly on it.
If so, and if a link crystal without target data is also found,
then the target data (simplified version without block info!)
is written to the first such link crystal.
And then the old target data is removed from the item.

Note: It should be safe to revert this commit to remove the compatibility
code after some time, maybe around 0.5.0 or 0.6.0.