Ender tools: Adjust the dig speeds. Use 2x durability in fast mode.
The dig speed is now set so that you can instant mine sandstone and
netherrack without any efficiency enchants in fast mode.
Stone and harder require Efficiency V, which gets and extra
speed boost (specifically checked for).
The Eff V speed boost is set to the minimun to allow instant mining
of iron blocks and other blocks that have hardness 5.0f, which
should be the highest on any "normal" blocks.