Add Public/Private mode checks to all items.
Note: The Ender Bucket can currently be used in machines regardless
of the privacy setting. Maybe I should add separate
fill() and drain() methods to do the work, and have the actual
fill() and drain() first do the privacy check, and then call the
internal methods to do the work?