Link Crystals/items: Add support for storing and using yaw and pitch rotations.
Storing a location including the yaw and pitch rotations is done with
a keybind combination: Alt + Shift + Toggle mode key.
Sneak + right click binding does not store the rotations.
If the rotation information is not present in the target tag,
then yaw and pitch of the teleported entity are not changed.