Ender Tools: Add binding the tool to inventories by sneak + right clicking.
Note: It is only possible to bind the tool to IInventory type blocks
or a vanilla Ender Chest by sneak + right clicking.
In-world locations need to be bound to the Link Crystals either separately
or by using the Alt + Shift + Toggle mode key combination.
(Sneak + right clicking on anything would interfere with the right
click block placing that comes next.)