Add model merging for items that have multiple textures.
All items now render completely, except the Ender Bucket fluid rendering.

Note: Since we just combine the models (all the BakedQuads are added
to the list), there is the possibility for z-fighting on overlapping
textures. So far I only saw it on the Ender Pickaxe on one side of one
pixel on the head of the pick, where the core(?) module overlaps the head.

In the future, it might be nice to do more intelligent module merging
so that we would either remove the inner layers, or somehow move the
outer layers a bit further out, so that they would be properly above the
inner layers.