Elite Ender Pearl: Fix going through blocks.
The cause was that the MovingObjectPosition of the block hit was sometimes
overridden by the entity hit with the riding player.

There is now a separate EntityThrowableEU class where the onUpdate() is
overriden with a version that ignores any entities that are in the "stack"
with the thrower of the pearl.