Mob Harness & player tick handler: Periodically re-add the AI task. Fix timer.
We now try to (re-)add the AI task every 2 seconds if it doesn't exist on
the bottom-most entity that the player is riding with the harness.
This fixes the need to remount the entity after using an Ender Porter while
riding it, for example.

Player tick handler timer was broken, the interval was one second only if
there was only one player on the server.