Change/clean-up/expand the keybinding handling.
The special case of spoofing Ctrl + mouse wheel as Shift + Toggle mode
is now gone.
Instead we now allow Alt, Shift and Control as modifiers with the
mouse wheel. The mouse wheel is now basically the same as the Toggle mode key.

Note: This unfortunately means that some keybind combinations had to change.
More or less the Control and Shift modifiers swapped place, so that
module selection is now done with Control, and an additional Shift
means the change is done backwards.
And similarly Control + Toggle mode combinations are now Shift + Toggle mode.

Also, the mouse wheel has a new "reverse" modifier.
That has to be checked for in addition to Shift for backwards actions.
(Unless we want to globally map Shift to the reverse condition?)