(Temporarily) fix the modifiers keys getting stuck if a GUI is opened
The KeyInputEvent doesn't fire inside GUIs, so the modifier keys
(the scrolling mode) got stuck if a GUI was opened while holding
an IKeyBound item and pressing Shift and/or Alt and/or Ctrl
and then opening a GUI, like the player inventory.

NOTE: This can probably be changed in 1.8 since it has apparently
has a separate key input event for GuiScreens. Although the
current method should also still work just fine.