Refactor Tool Workstation TileEntity and Container and Gui stuff
The TileEntity now only holds the tool and the module storage slots.
The installed modules are handled by an InventoryItem that takes
the tool as the container item.

There is also compatibility adjustment of installed modules,
if slot 0 is empty and there is a module installed in the
slot one after the last allowed slot.
This is because the installed modules previously started at slot 1.