Lazy Builder's Wand: Clean up/improve rendering, move start/end away from NBT.
Having the start and end corners in NBT caused problems when trying
to set the positions, because the change in NBT causes the event to fire
again immediately, so the positions would toggle each tick...

Also cleaned up the area rendering, and changed the walls and cube
modes' start and end positions to be the air block adjacent to
the existing targeted block.

And also added rendering the end corner of walls/cube in blue.