Lazy Builder's Wand: Change use behaviour with existing start position
The item now always uses the set/bound location, unless it is null.
Previously it would build into the pointed-to location, although
the rendered grid would be in the bound location.
Now it always only builds in the bound location, if one is set,
and if not, then it will build to the pointed-to location,
just like the rendering does.