Minor fixes to teleport code
There is an issue with client side ghost entities when teleporting
mobs that are riding other mobs.

Also when the player teleports from The End to the Overworld, the boss bar
sometimes gets stuck and also the fog distance becomes really low.
So there is some game state stuff that isn't properly handled atm.