Handy Bag: Fix an item loss bug on item pickup
If the Handy Bag GUI was open while the bag picked up items, then it
would create a new instance of the inventory and use that for inserting
the items. Then if any change was made in the open GUI, that would
overwrite the bag's NBT essentially deleting the picked up items in the
bag.