Builder's Wand: Refactor/fix/clean up the copy & paste template area handling
The wand now uses a separate "template metadata" file to read the area
size, which is used to find the opposite corner of the area.
Previously it was read from the copy mode's corner positions, but
that is a broken way to do it, since it will break if the copy mode
corners are changed after saving the area to template.
Also, the weird minimum/maximum corner code is now gone, and instead the
template size is saved relative to the start position (meaning that it
can also be negative).