Creation Station: Refactor crafting inventory stuff. Fixes JEI recipe derps.
Using the JEI recipe transfer GUI previously cleared the crafting
inventory stuff on the client side via onguiClosed/onContainerClosed
which caused the crafting t ofail on the client side, and the items on
cursor to be invisible unless the GUI was re-opened.

The crafting inventories aren't stored in the TE anymore, but are instead
wrapped when needed for the Container.

This refactor should also fix the missing public/private mode permission
checks for the crafting inventories. (untested atm though...)

Note: There is currently some sort of client sync issue when swapping
cards. Sometimes the crafting inventory doesn't properly get cleared
on the client side and the items from the previous card remain until
re-opening the GUI.