Fix some missing inventory wrapper method overrides. Fixes item loss issues.
The lockable inventory, and some of the wrappers were missing
isItemValidForSlot() overrides. Also the ASU was using the wrong
wrapper inventory. As a result, there was an item loss issue in the ASU
and the MSU when trying to click with items on a slot that was locked
into another item type.