1.11.2 port: Update some inventory and ItemStack things
This is just a small portion, and in no particular order. It's mostly
stackSize => setCount()/getCount() type of things (and nowhere near all of
them either), plus other ItemStack related things in the same classes.
Many of the base inventory classes (but not the wrappers) have now
also been udpated.