Don't sync the stored items for Handy Bag, Inventory Swapper or Memory Cards
The tag containing the stored items is not synced to the clients anymore.
Instead, a cached list of the stored ItemStack names and counts is stored
on the Memory Cards, to be used in the tooltip.

This should hopefully solve most of the issues with getting kicked out of
the server when nesting too many storage items inside the bags for
example.