Partially fix an issue of Handy Bag modes not being stored properly, closes #73
The mode will still be wiped from a fresh Card where it was just changed,
if either the Card is taken out of the bag immediately after, or items are
stored to the Card immediately after.

Those issue will hopefully be solved in the upcoming item-based inventory
overhaul, whenever that happens.